pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tpeairport"
version = "0.1.0"
description = "Scrape Taoyuan Airport flight forecast hourly people estimates (T1/T2)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "playwright>=1.44.0",
  "python-calamine>=0.6.2",
  "pytest>=8.0.0",
  "line-bot-sdk>=3.12.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

